+++ drizzle/0012_warm_nextwave.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `member` ADD `mastodon_account_id` text;--> statement-breakpoint
+ALTER TABLE `member` ADD `mastodon_access_token` text;